chore: reorganize pages import (#48)
