+++ src/lib/components/hall/components/CreateRoom.tsx
@@ -6,8 +6,8 @@ import { useState } from "react";
 
 import type { CreateRoomFormType } from "../types";
 import { CreateRoomFormSchema } from "../types";
-import SpokerInput from "lib/components/ui/SpokerInput";
-import SpokerWrapperGrid from "lib/components/ui/SpokerWrapperGrid";
+import SpokerInput from "lib/components/shared/SpokerInput";
+import SpokerWrapperGrid from "lib/components/shared/SpokerWrapperGrid";
 import { createRoom } from "lib/services/firebase/room";
 
 const CreateRoom = () => {
