+++ src/lib/components/layout/FullScreenLoading.tsx
@@ -1,7 +1,7 @@
 import type { LayoutProps } from "@chakra-ui/react";
 import { Grid } from "@chakra-ui/react";
 
-import SpokerLoading from "lib/components/ui/SpokerLoading";
+import SpokerLoading from "lib/components/shared/SpokerLoading";
 
 type FullScreenLoadingProps = {
   height?: LayoutProps["height"];
