+++ src/lib/components/room/PointWrapper.tsx
@@ -1,6 +1,6 @@
 import { Text } from "@chakra-ui/react";
 
-import Twemoji from "lib/components/ui/Twemoji";
+import Twemoji from "lib/components/shared/Twemoji";
 import type { HideLabelOptionsType } from "lib/constants/hideLabel";
 import { hideLabel } from "lib/constants/hideLabel";
