+++ src/pages/404.tsx
@@ -1,44 +1,3 @@
-import {
-  Box,
-  Button,
-  Heading,
-  Image,
-  Link as ChakraLink,
-  Text,
-  useColorMode,
-} from "@chakra-ui/react";
-import Link from "next/link";
-
-const Page404 = () => {
-  const { colorMode } = useColorMode();
-
-  return (
-    <>
-      <Box width={["100%", "70%", "60%", "60%"]} margin="0 auto">
-        <Image src="/404 Error-pana.svg" alt="error-404" />
-      </Box>
-      <Text textAlign="center" fontSize="xs">
-        <ChakraLink href="https://stories.freepik.com/web" isExternal>
-          Illustration by Freepik Stories
-        </ChakraLink>
-      </Text>
-
-      <Box marginY={4}>
-        <Heading textAlign="center">Page not Found.</Heading>
-
-        <Box textAlign="center" marginTop={4}>
-          <Text>It&apos;s Okay!</Text>
-          <Link href="/" passHref>
-            <Button
-              backgroundColor={colorMode === "light" ? "gray.300" : "teal.500"}
-            >
-              Let&apos;s Head Back
-            </Button>
-          </Link>
-        </Box>
-      </Box>
-    </>
-  );
-};
+import Page404 from "lib/pages/404";
 
 export default Page404;
